+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 All notable changes to this project will be documented in this file. This
 project adheres to [Semantic Versioning](http://semver.org/).
 
+## (next)
+- NEW: cf_notify as replacement for standard notify to avoid its refresh side effects
+
 ## 0.12.7 (2018-03-15)
 - CHANGED: to always prefer PSON serialization to avoid retries
 - CHANGED: moved syslog related stuff solely to cflogsink